[#5] Restructure qos mismatch error handling
