+++ rmw_iceoryx2_cxx/include/rmw_iceoryx2_cxx/impl/common/error.hpp
@@ -34,6 +34,7 @@ enum class PublisherError : uint8_t {
     INVARIANT_VIOLATION,
     SERVICE_NAME_CREATION_FAILURE,
     SERVICE_CREATION_FAILURE,
+    QOS_INCOMPATIBLE,
     PUBLISHER_CREATION_FAILURE,
     NOTIFIER_CREATION_FAILURE,
     LOAN_FAILURE,
@@ -45,6 +46,7 @@ enum class SubscriberError : uint8_t {
     INVARIANT_VIOLATION,
     SERVICE_NAME_CREATION_FAILURE,
     SERVICE_CREATION_FAILURE,
+    QOS_INCOMPATIBLE,
     SUBSCRIBER_CREATION_FAILURE,
     RECV_FAILURE,
     INVALID_PAYLOAD,
