+++ rmw_iceoryx2_cxx/src/rmw/publisher.cpp
@@ -18,8 +18,10 @@
 #include "rmw_iceoryx2_cxx/impl/common/ensure.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/error_message.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/log.hpp"
-#include "rmw_iceoryx2_cxx/impl/common/qos.hpp"
+#include "rmw_iceoryx2_cxx/impl/common/names.hpp"
+#include "rmw_iceoryx2_cxx/impl/common/qos_attributes.hpp"
 #include "rmw_iceoryx2_cxx/impl/message/introspection.hpp"
+#include "rmw_iceoryx2_cxx/impl/middleware/iceoryx2.hpp"
 #include "rmw_iceoryx2_cxx/impl/runtime/context.hpp"
 
 extern "C" {
@@ -55,6 +57,7 @@ rmw_publisher_t* rmw_create_publisher(const rmw_node_t* rmw_node,
     using ::rmw::iox2::ProfileKind;
     using ::rmw::iox2::ResolvedQos;
     using ::rmw::iox2::TryConvert;
+    using Iceoryx2 = ::rmw::iox2::Iceoryx2;
     using NodeImpl = ::rmw::iox2::Node;
     using PublisherImpl = ::rmw::iox2::Publisher;
     using ::rmw::iox2::unsafe_cast;
@@ -103,17 +106,29 @@ rmw_publisher_t* rmw_create_publisher(const rmw_node_t* rmw_node,
         RMW_IOX2_CHAIN_ERROR_MSG("failed to allocate memory for Publisher");
         return nullptr;
     } else {
-        if (!create_in_place<PublisherImpl>(
-                 publisher_impl.value(), *node_impl.value(), topic_name, type_support, resolved_qos.value())
-                 .has_value()) {
+        auto construct_result = create_in_place<PublisherImpl>(
+            publisher_impl.value(), *node_impl.value(), topic_name, type_support, resolved_qos.value());
+
+        if (!construct_result.has_value()) {
+            if (construct_result.error() == PublisherImpl::ErrorType::QOS_INCOMPATIBLE) {
+                auto details = node_impl.value()->iox2().lookup_service<Iceoryx2::ServiceType::Ipc>(
+                    ::rmw::iox2::names::topic(topic_name), Iceoryx2::MessagingPattern::PublishSubscribe);
+                if (!details.has_value()) {
+                    RMW_IOX2_CHAIN_ERROR_MSG_WITH_FORMAT_STRING(
+                        "QoS mismatch on '%s' (failed to look up service details)", topic_name);
+                }
+
+                ::rmw::iox2::chain_attribute_mismatch_error(
+                    resolved_qos.value(), details.value().static_details.attributes(), topic_name);
+            } else {
+                RMW_IOX2_CHAIN_ERROR_MSG("failed to construct Publisher");
+            }
             destruct<PublisherImpl>(publisher_impl.value());
             deallocate<PublisherImpl>(publisher_impl.value());
             rmw_publisher_free(rmw_publisher);
-            RMW_IOX2_CHAIN_ERROR_MSG("failed to construct Publisher");
             return nullptr;
-        } else {
-            rmw_publisher->data = publisher_impl.value();
         }
+        rmw_publisher->data = publisher_impl.value();
     }
 
     return rmw_publisher;
