+++ rmw_iceoryx2_cxx/src/rmw/subscription.cpp
@@ -18,8 +18,10 @@
 #include "rmw_iceoryx2_cxx/impl/common/ensure.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/error_message.hpp"
 #include "rmw_iceoryx2_cxx/impl/common/log.hpp"
-#include "rmw_iceoryx2_cxx/impl/common/qos.hpp"
+#include "rmw_iceoryx2_cxx/impl/common/names.hpp"
+#include "rmw_iceoryx2_cxx/impl/common/qos_attributes.hpp"
 #include "rmw_iceoryx2_cxx/impl/message/introspection.hpp"
+#include "rmw_iceoryx2_cxx/impl/middleware/iceoryx2.hpp"
 #include "rmw_iceoryx2_cxx/impl/runtime/context.hpp"
 #include "rmw_iceoryx2_cxx/impl/runtime/subscriber.hpp"
 
@@ -55,6 +57,7 @@ rmw_subscription_t* rmw_create_subscription(const rmw_node_t* rmw_node,
     using ::rmw::iox2::ProfileKind;
     using ::rmw::iox2::ResolvedQos;
     using ::rmw::iox2::TryConvert;
+    using Iceoryx2 = ::rmw::iox2::Iceoryx2;
     using NodeImpl = ::rmw::iox2::Node;
     using SubscriberImpl = ::rmw::iox2::Subscriber;
     using ::rmw::iox2::unsafe_cast;
@@ -103,17 +106,29 @@ rmw_subscription_t* rmw_create_subscription(const rmw_node_t* rmw_node,
         RMW_IOX2_CHAIN_ERROR_MSG("failed to allocate memory for Subscriber");
         return nullptr;
     } else {
-        if (!create_in_place<SubscriberImpl>(
-                 subscriber_impl.value(), *node_impl.value(), topic_name, type_support, resolved_qos.value())
-                 .has_value()) {
+        auto construct_result = create_in_place<SubscriberImpl>(
+            subscriber_impl.value(), *node_impl.value(), topic_name, type_support, resolved_qos.value());
+
+        if (!construct_result.has_value()) {
+            if (construct_result.error() == SubscriberImpl::ErrorType::QOS_INCOMPATIBLE) {
+                auto details = node_impl.value()->iox2().lookup_service<Iceoryx2::ServiceType::Ipc>(
+                    ::rmw::iox2::names::topic(topic_name), Iceoryx2::MessagingPattern::PublishSubscribe);
+                if (!details.has_value()) {
+                    RMW_IOX2_CHAIN_ERROR_MSG_WITH_FORMAT_STRING(
+                        "QoS mismatch on '%s' (failed to look up service details)", topic_name);
+                }
+
+                ::rmw::iox2::chain_attribute_mismatch_error(
+                    resolved_qos.value(), details.value().static_details.attributes(), topic_name);
+            } else {
+                RMW_IOX2_CHAIN_ERROR_MSG("failed to construct Subscriber");
+            }
             destruct<SubscriberImpl>(subscriber_impl.value());
             deallocate<SubscriberImpl>(subscriber_impl.value());
             rmw_subscription_free(rmw_subscription);
-            RMW_IOX2_CHAIN_ERROR_MSG("failed to construct Subscriber");
             return nullptr;
-        } else {
-            rmw_subscription->data = subscriber_impl.value();
         }
+        rmw_subscription->data = subscriber_impl.value();
     }
 
     return rmw_subscription;
